2015-02-23 Roman - FIX:SYS,Transport.Sockets,UI.{Vision,WebVision}: Some warnings about comparing signed and unsigned fix. - FIX:DAQ.MMS: RejectErrs and ConfirmErrs obtain fixed. - FIX:DAQ.Siemens: LibnoDave: Pointer to int cast fixed by replace the cast to long.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@2218 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/daq/Siemens/nodave.c
@@ -1898,7 +1898,7 @@ float DECL2 daveGetKGAt(daveConnection * dc,int pos) {
 
 float DECL2 daveGetKG(daveConnection * dc) {
     float f;
-    f=daveGetKGAt(dc, ((int)dc->resultPointer-(int)dc->_resultPointer));
+    f=daveGetKGAt(dc, ((long)dc->resultPointer-(long)dc->_resultPointer));
     dc->resultPointer+=4;
     return f;
 }
